requirements.txt
fastapi
uvicorn
openai
python-dotenv
autogen
faiss-cpu
requests
newspaper3k
beautifulsoup4
sentence-transformers
langchain
langgraph
langchain_community
langchain-openai
fpdf 
lxml_html_clean